Move interfaces out of file
