+++ src/utils/objectToWrapper.ts
@@ -1,0 +1,26 @@
+/**
+ * Copyright (C) 2021 PythonCoderAS
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+function objectToWrapper(input: { [k: string]: any }): Record<string, string> {
+  const obj: { [k: string]: string } = {};
+  for (const key in input) {
+    if (input.hasOwnProperty(key)) {
+      obj[key] = input[key]?.toString() || "";
+    }
+  }
+  return obj;
+}
