+++ src/utils/produceScreenshot/index.ts
@@ -18,59 +18,9 @@ import * as Buffer from "buffer";
 import { DateTime } from "luxon";
 import axios from "axios";
 
-import { Browser, devices } from "puppeteer";
-/* eslint-disable @typescript-eslint/no-var-requires -- Really hates it if I don't do this */
-const puppeteer = require("puppeteer");
-/* eslint-enable @typescript-eslint/no-var-requires */
-
-export let browser: Browser | null = null;
-
-export interface screeningTypeType {
-  G: string
-  S: string
-  E: string
-}
-
-export const screeningTypes: screeningTypeType = {
-  "G": "guest",
-  "S": "student",
-  "E": "employee"
-}
-
-export interface GenerateScreenshotParams {
-  firstName: string;
-  lastName: string;
-  email: string;
-  isVaxxed: boolean;
-  deviceName?: string;
-  type?: keyof screeningTypeType;
-}
-
-export interface SubmitParams {
-  Type: keyof screeningTypeType;
-  IsOther: boolean;
-  IsStudent: 0 | 1;
-  FirstName: string;
-  LastName: string;
-  Email: string;
-  State: string;
-  Location: string;
-  Floor: string | null;
-  Answer1: 0 | 1;
-  Answer2: 0 | 1;
-  Answer3?: 0 | 1;
-  Answer4?: 0 | 1;
-}
-
-function objectToWrapper(input: { [k: string]: any }): Record<string, string> {
-  const obj: { [k: string]: string } = {};
-  for (const key in input) {
-    if (input.hasOwnProperty(key)) {
-      obj[key] = input[key]?.toString() || "";
-    }
-  }
-  return obj;
-}
+import { devices } from "puppeteer";
+import { GenerateScreenshotParams, screeningTypes, SubmitParams } from "./interfaces";
+import { browser } from "./browser";
 
 export async function generateScreenshot(
   options: GenerateScreenshotParams
@@ -121,16 +71,7 @@ export async function generateScreenshot(
 
     screenshot = (await page.screenshot()) as Buffer;
   } finally {
-    // await page.close();
+    await page.close();
   }
   return screenshot;
-}
-
-export async function startupBrowser(): Promise<void> {
-  if (browser === null) {
-    browser = await puppeteer.launch({
-      headless: false,
-      // executablePath: "chromium-browser",
-    });
-  }
 }
